Add automated test for keyboard navigation
